Fixed error to send over maxlength text to youtube livechat

+++ example.py
@@ -2,9 +2,9 @@
 
 from time import sleep
 
-from youtubechat import YoutubeLiveChat, get_live_chat_id_for_broadcast_id,get_live_chat_id_for_stream_now
+from youtubechat import YoutubeLiveChat, get_live_chat_id_for_broadcast_id, get_live_chat_id_for_stream_now
 
-broadcast_id = 'BRHNUf8gTQE'
+broadcast_id = 'zDWUNqe3alo'
 
 livechat_id = get_live_chat_id_for_broadcast_id(broadcast_id,"oauth_creds")
 print(livechat_id)
@@ -14,9 +14,9 @@ chat_obj = YoutubeLiveChat("oauth_creds", [livechat_id])
 def respond(msgs, chatid):
     for msg in msgs:
         print(msg)
-        msg.delete()
-        chat_obj.send_message("RESPONSE!", chatid)
-
+        if msg.message_text[0] is '!':
+            chat_obj.send_message("DDOKDDOK : "+msg.message_text, chatid)
+            msg.delete()
 
 try:
     chat_obj.start()
